+++ src/Proxy/CircularStringProxy.php
@@ -223,7 +223,7 @@ class CircularStringProxy extends CircularString implements ProxyInterface
             $this->load();
         }
 
-        return $this->proxyGeometry->withAddedPoints($points);
+        return $this->proxyGeometry->withAddedPoints(...$points);
     }
 
     public function coordinateDimension(): int