+++ src/Proxy/MultiCurveProxy.php
@@ -214,7 +214,7 @@ class MultiCurveProxy extends MultiCurve implements ProxyInterface
             $this->load();
         }
 
-        return $this->proxyGeometry->withAddedGeometries($geometries);
+        return $this->proxyGeometry->withAddedGeometries(...$geometries);
     }
 
     public function coordinateDimension(): int