+++ src/Proxy/MultiPointProxy.php
@@ -214,7 +214,7 @@ class MultiPointProxy extends MultiPoint implements ProxyInterface
             $this->load();
         }
 
-        return $this->proxyGeometry->withAddedGeometries($geometries);
+        return $this->proxyGeometry->withAddedGeometries(...$geometries);
     }
 
     public function coordinateDimension(): int